FIX: resetting av utfylte felter i korrigering
