+++ src/app/sider/innsending/innsendingRoutes.tsx
@@ -9,18 +9,12 @@ import { connect } from 'react-redux';
 import { Dispatch } from 'redux';
 import { InnsendingActions } from '../../actions/innsending';
 import { InnsendingState, Innsendingstyper } from '../../types/innsending';
-import {
-  Meldekort,
-  MeldekortDag,
-  Meldekortdetaljer,
-  Sporsmal,
-} from '../../types/meldekort';
+import { Meldekort, Meldekortdetaljer } from '../../types/meldekort';
 import { Redirect, Route, RouteComponentProps, Switch } from 'react-router-dom';
 import { RootState } from '../../store/configureStore';
 import { Sporsmal as Spm } from './1-sporsmalsside/sporsmal/sporsmalConfig';
 import { MeldekortdetaljerActions } from '../../actions/meldekortdetaljer';
 import { UtfyltDag } from './2-utfyllingsside/utfylling/utfyltDagConfig';
-import { hentNorskeUkedager } from '../../utils/ukedager';
 import { RouterState } from 'connected-react-router';
 
 interface MapStateToProps {
@@ -58,89 +52,8 @@ class InnsendingRoutes extends React.Component<InnsendingRoutesProps> {
     }
   };
 
-  settSporsmalOgUtfyllingHvisKorrigering = () => {
-    const {
-      innsending,
-      oppdaterSporsmalsobjekter,
-      oppdaterUtfylteDager,
-    } = this.props;
-    if (innsending.innsendingstype === Innsendingstyper.korrigering) {
-      const konverterteSporsmalsobjekter = this.konverterMeldekortdetaljerSporsmalTilInnsendingSporsmal(
-        this.props.meldekortdetaljer.sporsmal,
-        innsending.sporsmalsobjekter
-      );
-      const konverterteUtfylteDager = this.konverterMeldekortdetaljerMeldekortDagerTilInnsendingUtfylteDager(
-        this.props.meldekortdetaljer.sporsmal.meldekortDager,
-        innsending.utfylteDager
-      );
-      oppdaterSporsmalsobjekter(konverterteSporsmalsobjekter);
-      oppdaterUtfylteDager(konverterteUtfylteDager);
-    }
-  };
-
-  returnerListeMedMeldekortdetaljerSporsmal = (
-    mkdetaljerSporsmal: Sporsmal
-  ) => {
-    return [
-      { kategori: 'arbeid', checked: mkdetaljerSporsmal.arbeidet },
-      { kategori: 'aktivitetArbeid', checked: mkdetaljerSporsmal.kurs },
-      { kategori: 'forhindret', checked: mkdetaljerSporsmal.syk },
-      { kategori: 'ferieFravar', checked: mkdetaljerSporsmal.annetFravaer },
-      { kategori: 'registrert', checked: mkdetaljerSporsmal.arbeidssoker },
-    ];
-  };
-
-  settCheckedBasertPaBoolean = (
-    kategoritekst: string,
-    sporsmalValg: boolean
-  ) => {
-    return sporsmalValg ? kategoritekst + '.ja' : kategoritekst + '.nei';
-  };
-
-  konverterMeldekortdetaljerMeldekortDagerTilInnsendingUtfylteDager = (
-    meldekortDager: MeldekortDag[],
-    utfylteDager: UtfyltDag[]
-  ) => {
-    const ukedagerSomListe = hentNorskeUkedager();
-    return utfylteDager.map((utfyltDag, index) => {
-      return {
-        ...utfyltDag,
-        uke: index < 7 ? 1 : 2,
-        dag: ukedagerSomListe[index % 7].trim(),
-        syk: meldekortDager[index].syk,
-        arbeidetTimer: meldekortDager[index].arbeidetTimerSum.toString(),
-        annetFravaer: meldekortDager[index].annetFravaer,
-        kurs: meldekortDager[index].kurs,
-      };
-    });
-  };
-
-  konverterMeldekortdetaljerSporsmalTilInnsendingSporsmal = (
-    mkdetaljerSporsmal: Sporsmal,
-    innsendingSporsmal: Spm[]
-  ): Spm[] => {
-    const listeMedSporsmal = mkdetaljerSporsmal!
-      ? this.returnerListeMedMeldekortdetaljerSporsmal(mkdetaljerSporsmal)
-      : [];
-    return innsendingSporsmal.map(spm => {
-      for (let i = 0; i < listeMedSporsmal.length; i++) {
-        if (spm.kategori === listeMedSporsmal[i].kategori) {
-          return {
-            ...spm,
-            checked: this.settCheckedBasertPaBoolean(
-              spm.kategori,
-              listeMedSporsmal[i].checked
-            ),
-          };
-        }
-      }
-      return { ...spm };
-    });
-  };
-
   componentDidMount() {
     this.settMeldekortIdBasertPaInnsendingstype();
-    this.settSporsmalOgUtfyllingHvisKorrigering();
   }
 
   render() {
