+++ src/app/utils/meldekortUtils.tsx
@@ -1,11 +1,17 @@
 import {
   KortStatus,
   Meldekort,
+  MeldekortDag,
+  Meldekortdetaljer,
   MeldekortRad,
   SendtMeldekort,
+  Sporsmal,
 } from '../types/meldekort';
-import { Innsendingstyper } from '../types/innsending';
+import { InnsendingState, Innsendingstyper } from '../types/innsending';
 import { hentDatoPeriode, hentUkePeriode } from './dates';
+import { Sporsmal as Spm } from '../sider/innsending/1-sporsmalsside/sporsmal/sporsmalConfig';
+import { UtfyltDag } from '../sider/innsending/2-utfyllingsside/utfylling/utfyltDagConfig';
+import { hentNorskeUkedager } from './ukedager';
 
 export const erMeldekortSendtInnTidligere = (
   meldekort: Meldekort,
